add GameObserver, play_from, and RandomOpening game variant
Introduces infrastructure for mid-game board injection and per-move
hooks, and uses it to support random opening positions in evaluation.

- Add `GameObserver` trait with `on_move` hook (fires before placement,
  can break early); add `NoOpObserver` ZST
- Refactor `Play` trait: `play_from(&mut Board, observer, ...)` is now
  the required method; `play()` becomes a provided default using an
  empty board and `NoOpObserver`
- Add `pub run_from` shared game loop used by all game variants
- `Freestyle` delegates to `run_from`; removes bespoke `run_internal`
- Add `RandomOpening { moves: u32 }` game variant: pre-places N random
  stones on the board before handing off to `run_from`
- Refactor `MinimaxFitness`: replace bespoke `play_game` loop with
  `game: Game` field + `MinimaxObserver` (move limit + per-move scoring)
- Change `MatchResult::turn_count` from `u32` to `usize` to match
  `Board::move_count()`
- Add `--opening-moves N` CLI arg to `evolve`; selects `RandomOpening`
  or `Freestyle` for both tournament and minimax evaluators

+++ src/match_result.rs
@@ -6,7 +6,7 @@ pub struct MatchResult {
     outcome: Outcome,
     black_label: String,
     white_label: String,
-    turn_count: u32,
+    turn_count: usize,
     board_state: String,
 }
 
@@ -16,7 +16,7 @@ impl MatchResult {
         outcome: Outcome,
         black_label: String,
         white_label: String,
-        turn_count: u32,
+        turn_count: usize,
         board_state: String,
     ) -> Self {
         Self {
@@ -44,7 +44,7 @@ impl MatchResult {
     }
 
     #[must_use]
-    pub fn turn_count(&self) -> u32 {
+    pub fn turn_count(&self) -> usize {
         self.turn_count
     }
